Moving binary incompatible change out into FoldableSyntaxBinCompat1
