+++ core/src/main/scala/cats/syntax/all.scala
@@ -84,3 +84,5 @@ trait AllSyntaxBinCompat4
     with ParallelApplySyntax
     with FoldableSyntaxBinCompat0
     with ReducibleSyntaxBinCompat0
+
+trait AllSyntaxBinCompat5 extends FoldableSyntaxBinCompat1
